A few misc changes
Moved Image to its own class, split up edit_X_() methods and removed
Channel#getParent()
